profileUpdation

+++ client/src/components/Logout.js
@@ -9,6 +9,8 @@ function Logout() {
     const navigate=useNavigate();
     const dispatch=useDispatch();
     const logout = () => {
+        let flag=window.confirm("want to logout?");
+        if(!flag) return;
         localStorage.removeItem("token");
         dispatch(remove());
         navigate("/");
@@ -16,8 +18,8 @@ function Logout() {
     const userData=useSelector(state=>state.user);
 
     return (userData.fullName) ?
-        (<Link onClick={logout} ><span className='hover:text-cyan-500 text-xl duration-500 mx-4 my-4 md:mx-6 md:my:0' >Logout</span></Link>)
-        : <Link className={`hover:text-cyan-500 text-xl duration-500 mx-4 my-4 md:mx-6 md:my:0`} to="/register">Signup/Login</Link>
+        (<Link onClick={logout} className='hover:text-cyan-500 text-xl duration-500 mx-4 my-2 md:mx-6 md:my:0' ><span  >Logout</span></Link>)
+        : <Link  className='hover:text-cyan-500 text-xl duration-500 mx-4 my-2 md:mx-6 md:my:0' to="/register">Signup/Login</Link>
 }
 
 export default Logout
